EY-2646: Flyttar OMS-spesifikt vedlegg ut til OMS-pakka
